refactor(crdt): move IPC handler registration into ipc/crdt-handlers.ts
Satisfies architecture boundary rule: IPC registration must live in
apps/desktop/src/main/ipc/ not sync/. Moves registerCrdtIpcHandlers
+ _resetCrdtIpcHandlersForTests out of crdt-provider.ts into a new
ipc/crdt-handlers.ts module, adds it to the sync-boundary exemption
list alongside sync-core-handlers.ts et al.

Also adds CrdtProvider.validateNoteForCrdt() so the IPC handler does
not directly hit @main/database/queries/notes (another boundary rule).
